fix: properly overwrite and start new evaluations
previously after pressing an operator button the numbers displayed
were added on to the string, now it properly overwrites the
previous input.

the equal button was also not working as expected as after pressing
it and pressing an new number which should start a new evaluation
instead it continued and did the operation on top of the previous
result. now it is acting as expected.